src/main/java/teste/gestaovendas/service/ClienteService.java
package teste.gestaovendas.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import teste.gestaovendas.model.Cliente;
import teste.gestaovendas.repository.ClienteRepository;
import teste.gestaovendas.repository.PedidoRepository;

import java.util.List;
import java.util.Optional;

@Service
public class ClienteService {

    @Autowired
    private ClienteRepository clienteRepository;

    private PedidoRepository pedidoRepository;

    public List<Cliente> listarTodos() {
        return clienteRepository.findAll();
    }

    public Optional<Cliente> buscarPorId(Long id) {
        return clienteRepository.findById(id);
    }

    public Cliente salvar(Cliente cliente) {
        return clienteRepository.save(cliente);
    }

    public void deletarPorId(Long id) {
        clienteRepository.deleteById(id);
    }

    public Cliente atualizar(Long id, Cliente clienteAtualizado) {
        return clienteRepository.findById(id).map(cliente -> {
            cliente.setNome(clienteAtualizado.getNome());
            cliente.setLimiteCompra(clienteAtualizado.getLimiteCompra());
            cliente.setDiaFechamentoFatura(clienteAtualizado.getDiaFechamentoFatura());
            return clienteRepository.save(cliente);
        }).orElseThrow(() -> new RuntimeException("Cliente não encontrado"));
    }
}